Organize directory structure and naming rule
