+++ src/main/ts/api/folder-test.ts
@@ -4,10 +4,10 @@ import {describe, it} from 'mocha';
 import {ButtonController} from '../controller/button';
 import {FolderController} from '../controller/folder';
 import {InputBindingController} from '../controller/input-binding';
-import {NumberTextController} from '../controller/input/number-text';
 import {MonitorBindingController} from '../controller/monitor-binding';
-import {SingleLogMonitorController} from '../controller/monitor/single-log';
 import {SeparatorController} from '../controller/separator';
+import {NumberTextController} from '../controller/value/number-text';
+import {SingleLogMonitorController} from '../controller/value/single-log';
 import {TestUtil} from '../misc/test-util';
 import {Color} from '../model/color';
 import {ViewModel} from '../model/view-model';
