+++ src/main/ts/controller/value/a-palette.ts
