+++ src/main/ts/controller/value/color-component-texts-test.ts
