+++ src/main/ts/controller/value/color-component-texts.ts
