+++ src/main/ts/controller/value/color-picker.ts
