+++ src/main/ts/controller/value/color-swatch-test.ts
