+++ src/main/ts/controller/value/color-swatch-text-test.ts
