+++ src/main/ts/controller/value/color-swatch-text.ts
