+++ src/main/ts/controller/value/color-swatch.ts
