+++ src/main/ts/controller/value/h-palette-test.ts
