+++ src/main/ts/controller/value/h-palette.ts
