+++ src/main/ts/controller/value/number-text-test.ts
