+++ src/main/ts/controller/value/number-text.ts
