+++ src/main/ts/controller/value/point-2d-pad-text.ts
