+++ src/main/ts/controller/value/point-2d-text-test.ts
