+++ src/main/ts/controller/value/single-log.ts
@@ -2,7 +2,7 @@ import {Formatter} from '../../formatter/formatter';
 import {Buffer, BufferedValue} from '../../model/buffered-value';
 import {ViewModel} from '../../model/view-model';
 import {SingleLogView} from '../../view/value/single-log';
-import {ValueController} from '../input/value';
+import {ValueController} from '../value/value';
 
 interface Config<T> {
 	formatter: Formatter<T>;
