+++ src/main/ts/controller/value/slider-test.ts
