+++ src/main/ts/controller/value/slider-text-test.ts
