+++ src/main/ts/controller/value/slider-text.ts
