+++ src/main/ts/controller/value/slider.ts
