+++ src/main/ts/controller/value/sv-palette-test.ts
