+++ src/main/ts/controller/value/sv-palette.ts
