+++ src/main/ts/controller/value/text-test.ts
