+++ src/main/ts/pane/input-number-test.ts
@@ -4,9 +4,9 @@ import {describe as context, describe, it} from 'mocha';
 import {Constraint} from '../constraint/constraint';
 import {StepConstraint} from '../constraint/step';
 import {ConstraintUtil} from '../constraint/util';
-import {ListController} from '../controller/input/list';
-import {SliderTextController} from '../controller/input/slider-text';
-import {TextController} from '../controller/input/text';
+import {ListController} from '../controller/value/list';
+import {SliderTextController} from '../controller/value/slider-text';
+import {TextController} from '../controller/value/text';
 import {TestUtil} from '../misc/test-util';
 import {TypeUtil} from '../misc/type-util';
 import {Value} from '../model/value';
