+++ src/main/ts/pane/monitor-number-test.ts
@@ -2,9 +2,9 @@ import {assert} from 'chai';
 import {describe as context, describe, it} from 'mocha';
 
 import {MonitorBinding} from '../binding/monitor';
-import {GraphLogController} from '../controller/monitor/graph-log';
-import {MultiLogController} from '../controller/monitor/multi-log';
-import {SingleLogMonitorController} from '../controller/monitor/single-log';
+import {GraphLogController} from '../controller/value/graph-log';
+import {MultiLogController} from '../controller/value/multi-log';
+import {SingleLogMonitorController} from '../controller/value/single-log';
 import {TestUtil} from '../misc/test-util';
 import {IntervalTicker} from '../misc/ticker/interval';
 import {PlainTweakpane} from './plain-tweakpane';
