+++ src/main/ts/plugin/input-bindings/boolean.ts
@@ -3,8 +3,8 @@ import {CompositeConstraint} from '../../constraint/composite';
 import {Constraint} from '../../constraint/constraint';
 import {ListConstraint} from '../../constraint/list';
 import {ConstraintUtil} from '../../constraint/util';
-import {CheckboxController} from '../../controller/input/checkbox';
-import {ListController} from '../../controller/input/list';
+import {CheckboxController} from '../../controller/value/checkbox';
+import {ListController} from '../../controller/value/list';
 import * as BooleanConverter from '../../converter/boolean';
 import {Value} from '../../model/value';
 import {ViewModel} from '../../model/view-model';
