+++ src/main/ts/plugin/input-bindings/color-number.ts
@@ -1,5 +1,5 @@
 import {InputParams} from '../../api/types';
-import {ColorSwatchTextController} from '../../controller/input/color-swatch-text';
+import {ColorSwatchTextController} from '../../controller/value/color-swatch-text';
 import * as ColorConverter from '../../converter/color';
 import {ColorFormatter} from '../../formatter/color';
 import {Color} from '../../model/color';
