+++ src/main/ts/plugin/input-bindings/number.ts
@@ -5,9 +5,9 @@ import {ListConstraint} from '../../constraint/list';
 import {RangeConstraint} from '../../constraint/range';
 import {StepConstraint} from '../../constraint/step';
 import {ConstraintUtil} from '../../constraint/util';
-import {ListController} from '../../controller/input/list';
-import {NumberTextController} from '../../controller/input/number-text';
-import {SliderTextController} from '../../controller/input/slider-text';
+import {ListController} from '../../controller/value/list';
+import {NumberTextController} from '../../controller/value/number-text';
+import {SliderTextController} from '../../controller/value/slider-text';
 import * as NumberConverter from '../../converter/number';
 import {NumberFormatter} from '../../formatter/number';
 import {TypeUtil} from '../../misc/type-util';
