+++ src/main/ts/plugin/input-bindings/string.ts
@@ -3,8 +3,8 @@ import {CompositeConstraint} from '../../constraint/composite';
 import {Constraint} from '../../constraint/constraint';
 import {ListConstraint} from '../../constraint/list';
 import {ConstraintUtil} from '../../constraint/util';
-import {ListController} from '../../controller/input/list';
-import {TextController} from '../../controller/input/text';
+import {ListController} from '../../controller/value/list';
+import {TextController} from '../../controller/value/text';
 import * as StringConverter from '../../converter/string';
 import {StringFormatter} from '../../formatter/string';
 import {Value} from '../../model/value';
