+++ src/main/ts/plugin/monitor-binding.ts
@@ -1,7 +1,7 @@
 import {MonitorParams} from '../api/types';
 import {MonitorBinding} from '../binding/monitor';
-import {ValueController} from '../controller/input/value';
 import {MonitorBindingController} from '../controller/monitor-binding';
+import {ValueController} from '../controller/value/value';
 import {Constants} from '../misc/constants';
 import {IntervalTicker} from '../misc/ticker/interval';
 import {ManualTicker} from '../misc/ticker/manual';
